Refactor icekit.admin* and icekit.utils.admin.* to icekit.admin_tools.*. Previous classes/functions are deprecated.
