+++ docs/customising/content-plugins.rst
@@ -94,7 +94,7 @@ models, with two mixin classes:
 
     # admin.py
 
-    from icekit.admin_mixins import FluentLayoutsMixin
+    from icekit.admin_tools.mixins import FluentLayoutsMixin
 
     class MyModelAdmin(FluentLayoutsMixin, MyModelAdminBase):
         ...
