+++ icekit/abstract_models.py
@@ -9,11 +9,14 @@ from django.template.loader import get_template
 from django.utils import encoding, timezone
 from django.utils.translation import ugettext_lazy as _
 from fluent_contents.analyzer import get_template_placeholder_data
-from fluent_contents.extensions import ContentPlugin
-from fluent_contents.models import ContentItem
-from icekit import appsettings
+from icekit.admin_tools.filters import ChildModelFilter
+
+
 
 from . import fields, plugins
+
+ChildModelFilter
+
 
 
 class AbstractBaseModel(models.Model):
