+++ icekit/admin_tools/__init__.py
@@ -1,0 +1,4 @@
+"""
+This module is named icekit.admin_tools in order not to clash with the
+existing icekit.admin, which registers admins for icekit.models.
+"""
