+++ icekit/plugins/slideshow/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from fluent_contents.admin import PlaceholderFieldAdmin
 
-from icekit.admin import ICEkitContentsAdmin
+from icekit.admin_tools.mixins import ICEkitContentsAdmin
 
 from . import models
