+++ icekit/utils/admin/__init__.py
@@ -1,0 +1,3 @@
+import warnings
+warnings.warn("the icekit.utils.admin module is deprecated. Use icekit.admin_tools.* instead.", DeprecationWarning,
+              stacklevel=2)
